Improvements on Male Corn data
MaleCorn (the 1st version of it) and male placement corn where to variables/columns containing the same data at different times in our data collection. In order to use this data I needed to regroup everything in a single colum, the new MaleCorn. Here are some of the last steps in my process to finish treating this part of the data