Cleaning of Dyad Distance in Process
I finished cleaning Dyad Distance by creating a new variable called Distance. I will stop for now since I think they are a few mistakes in it. I will continue the cleaning and consider later on if I have to keep this varaible or not after checking if the code works