Cleaning of the Remaing NA's
I cleand the NA's of each variables and it seems that there is none left. Neverthelesse, I will have to recheck these codes to see that everything was implemented correctly and that it they are no mistakes. Next time I will start rereading and creating new variables for my statistical part